Added asserts after ASN1 message encoding to detect buffer overflow or other encoding issues. Increase RRC buffer size for RRCConnectionReconfiguration messages.
git-svn-id: http://svn.eurecom.fr/openair4G/trunk@4926 818b1a75-f10b-46b9-bf7c-635c3b92a50f

+++ openair2/RRC/LITE/MESSAGES/asn1_msg.h
@@ -167,26 +167,27 @@ uint8_t do_RRCConnectionSetup(uint8_t Mod_id,
 @param cba_rnti RNTI for the cba transmission 
 @returns Size of encoded bit stream in bytes*/
 
-uint8_t do_RRCConnectionReconfiguration(uint8_t                           Mod_id,
-                                        uint8_t                          *buffer,
-                                        uint8_t                           UE_id,
-                                        uint8_t                           Transaction_id,
-                                        SRB_ToAddModList_t                *SRB_list,
-                                        DRB_ToAddModList_t                *DRB_list,
-                                        DRB_ToReleaseList_t               *DRB_list2,
-                                        struct SPS_Config                 *sps_Config,
-                                        struct PhysicalConfigDedicated    *physicalConfigDedicated,
-                                        MeasObjectToAddModList_t          *MeasObj_list,
-                                        ReportConfigToAddModList_t        *ReportConfig_list,
-                                        QuantityConfig_t                  *quantityConfig,
-                                        MeasIdToAddModList_t              *MeasId_list,
-                                        MAC_MainConfig_t                  *mac_MainConfig,
-                                        MeasGapConfig_t                   *measGapConfig,
-                                        MobilityControlInfo_t             *mobilityInfo,
-                                        struct MeasConfig__speedStatePars *speedStatePars,
-                                        RSRP_Range_t                      *rsrp,
-                                        C_RNTI_t                          *cba_rnti,
-                                        struct RRCConnectionReconfiguration_r8_IEs__dedicatedInfoNASList *dedicatedInfoNASList);
+uint16_t do_RRCConnectionReconfiguration(uint8_t                             Mod_id,
+                                         uint8_t                            *buffer,
+                                         uint8_t                             UE_id,
+                                         uint8_t                             Transaction_id,
+                                         SRB_ToAddModList_t                 *SRB_list,
+                                         DRB_ToAddModList_t                 *DRB_list,
+                                         DRB_ToReleaseList_t                *DRB_list2,
+                                         struct SPS_Config                  *sps_Config,
+                                         struct PhysicalConfigDedicated     *physicalConfigDedicated,
+                                         MeasObjectToAddModList_t           *MeasObj_list,
+                                         ReportConfigToAddModList_t         *ReportConfig_list,
+                                         QuantityConfig_t                   *quantityConfig,
+                                         MeasIdToAddModList_t               *MeasId_list,
+                                         MAC_MainConfig_t                   *mac_MainConfig,
+                                         MeasGapConfig_t                    *measGapConfig,
+                                         MobilityControlInfo_t              *mobilityInfo,
+                                         struct MeasConfig__speedStatePars  *speedStatePars,
+                                         RSRP_Range_t                       *rsrp,
+                                         C_RNTI_t                           *cba_rnti,
+                                         struct RRCConnectionReconfiguration_r8_IEs__dedicatedInfoNASList
+                                                                            *dedicatedInfoNASList);
 
 /***
  * \brief Generate an MCCH-Message (eNB). This routine configures MBSFNAreaConfiguration (PMCH-InfoList and Subframe Allocation for MBMS data)
